# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AuthResponse" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DevLoginRequest" : {
        "properties" : {
          "secret" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "secret" ],
        "type" : "object"
      },
...
      "SocialLoginRequest" : {
        "properties" : {
          "token" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "token" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "음성 기반 요리 어시스턴트 CookPilot 백엔드 API. 모든 라우트는 /api/v1 아래에 있고, 에러는 RFC-7807 ProblemDetail 형식이다.",
    "title" : "CookPilot Backend API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/dev" : {
      "post" : {
        "operationId" : "loginAsDeveloper",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/{provider}" : {
      "post" : {
        "operationId" : "login",
        "parameters" : [ {
          "in" : "path",
          "name" : "provider",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SocialLoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}